backend/requirements.txt
# === Core web stack ===
fastapi
uvicorn[standard]
gunicorn
pydantic>=2
pydantic-settings
python-dotenv
requests
python-multipart
aiofiles

# === AI / LangChain ===
langchain>=0.2.12
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
google-generativeai>=0.3.0

# === Vector DB ===
qdrant-client>=1.9.0

# === Embeddings (Jina AI - Primary) ===
# Jina uses standard requests library (already included above)

# === Embeddings (Optional - install only if needed) ===
# Uncomment the provider you want to use:

# Cohere (if using EMBEDDING_PROVIDER=cohere)
# cohere

# Voyage AI (if using EMBEDDING_PROVIDER=voyage)
# voyageai

# HuggingFace (legacy - not recommended)
# huggingface-hub==0.29.0

# === PDF Processing ===
PyPDF2